requirements.txt
# ────────────────────────────────────────────────────────────
# requirements.txt - 노이즈 스캐너 의존성 패키지 (Gemini 버전)
# 설치: pip install -r requirements.txt
# ────────────────────────────────────────────────────────────

# ── AI / LLM (Gemini 전환) ──────────────────────────────────
google-generativeai>=0.4.0     # Google Gemini API 클라이언트

# ── 텔레그램 봇 ───────────────────────────────────────────
python-telegram-bot>=21.0.0    # 텔레그램 봇 (async v20+)

# ── 데이터 수집 ───────────────────────────────────────────
aiohttp>=3.9.0                  # 비동기 HTTP 클라이언트
feedparser>=6.0.0               # RSS/Atom 피드 파서
beautifulsoup4>=4.12.0          # HTML 파싱
lxml>=5.0.0                     # BeautifulSoup 고속 파서 백엔드
yfinance>=0.2.40                # Yahoo Finance 데이터 (주가·뉴스)

# ── 유틸리티 ──────────────────────────────────────────────
python-dotenv>=1.0.0            # .env 파일 로드